+++ src/microbe_stage/editor/MicrobeEditorReportComponent.cs
@@ -411,7 +411,7 @@ public class MicrobeEditorReportComponent : EditorComponentBase<IEditorReportDat
     {
         Editor.SendAutoEvoResultsToReportComponent();
         UpdateTimeIndicator(Editor.CurrentGame.GameWorld.TotalPassedTime);
-        UpdateGlucoseReduction((float)Editor.CurrentGame.GameWorld.WorldSettings.GlucoseDecay);
+        UpdateGlucoseReduction(Editor.CurrentGame.GameWorld.WorldSettings.GlucoseDecay);
         UpdateTimeline(Editor.SelectedPatch);
         UpdateReportTabPatchSelector();
     }
